src/requirements.txt
# =============================================================================
# Medical Record Inspector - Python Dependencies
# =============================================================================
# Generated: 2026-03-19
# Python Version: 3.9+
# =============================================================================

# -----------------------------------------------------------------------------
# Core Web Framework
# -----------------------------------------------------------------------------
fastapi==0.109.0              # Web framework for API endpoints
uvicorn[standard]==0.27.0     # ASGI server for running FastAPI
pydantic==2.5.3               # Data validation using Python type annotations
pydantic-settings==2.12.0     # Settings management with Pydantic
pydantic_core==2.14.6         # Pydantic core components (auto-installed)

# -----------------------------------------------------------------------------
# HTTP Client (LLM API Calls)
# -----------------------------------------------------------------------------
httpx==0.26.0                 # Async HTTP client for LLM API calls
httpx-sse==0.4.0              # Server-sent events support for httpx

# -----------------------------------------------------------------------------
# Environment Variables
# -----------------------------------------------------------------------------
python-dotenv==1.0.0          # Load environment variables from .env files

# -----------------------------------------------------------------------------
# File Parsing (Document Processing)
# -----------------------------------------------------------------------------
python-docx==1.1.0            # DOCX file parsing
pypdf==5.8.0                  # PDF text extraction
pypdfium2==4.30.0             # PDF rendering backend (auto-installed)

# -----------------------------------------------------------------------------
# PDF Generation
# -----------------------------------------------------------------------------
reportlab==4.4.2              # PDF document generation

# -----------------------------------------------------------------------------
# Logging
# -----------------------------------------------------------------------------
structlog==24.1.0             # Structured logging

# -----------------------------------------------------------------------------
# CORS Support
# -----------------------------------------------------------------------------
python-cors==1.0.0            # CORS middleware for FastAPI

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest==9.0.2                 # Testing framework
pytest-asyncio==1.3.0         # Async testing support
pytest-cov==7.0.0             # Coverage reporting

# -----------------------------------------------------------------------------
# Optional: Unofficial LLM API Clients
# -----------------------------------------------------------------------------
# openai==1.10.0              # OpenAI API (uncomment if using OpenAI)
# dashscope==1.14.0           # Tongyi Qianwen API (uncomment if using)
# aistudio-sdk==0.3.8         # ERNIE Bot API (uncomment if using ERNIE Bot)

# -----------------------------------------------------------------------------
# Development & Utilities
# -----------------------------------------------------------------------------
# WeasyPrint not included - requires system dependencies (Cairo)
# For PDF export without system dependencies, use reportlab instead
# weasyprint==62.0            # HTML to PDF conversion (requires Cairo)
# =============================================================================

